requirements.txt
kivy[base]==2.2.1
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.6.0
kivymd==1.1.1
openai==1.12.0
python-dotenv==1.0.0 